refactor(engine): extract spill/tempfile.rs (SPL-3 #2325)
Move SpillBackend, ReadWriteSeek, and open_backend from spill.rs into a
new spill/tempfile.rs submodule. The tempfile-backed storage primitives
are self-contained (no back-references to SpillableReorderBuffer or
related types) and form a coherent disk-facing surface that the rest of
the spill layer consumes through a narrow internal API.

Pure code movement: no logic changes, no public API change. Visibility
narrowed to pub(super) since these items were previously module-private
and only consumed inside spill.rs. Tests still reach the symbols via the
same call chain through SpillableReorderBuffer.
